+++ Src/Public/Invoke-AsBuiltReport.Microsoft.Windows.ps1
@@ -5,7 +5,7 @@ function Invoke-AsBuiltReport.Microsoft.Windows {
     .DESCRIPTION
         Documents the configuration of Microsoft Windows Server in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.5.2
+        Version:        0.5.3
         Author:         Andrew Ramsay
         Editor:         Jonathan Colon
         Twitter:        @asbuiltreport
@@ -22,20 +22,20 @@ function Invoke-AsBuiltReport.Microsoft.Windows {
         [PSCredential] $Credential
     )
 
-    Write-PScriboMessage -IsWarning "Please refer to the AsBuiltReport.Microsoft.Windows github website for more detailed information about this project."
-    Write-PScriboMessage -IsWarning "Do not forget to update your report configuration file after each new release."
-    Write-PScriboMessage -IsWarning "Documentation: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.Windows"
-    Write-PScriboMessage -IsWarning "Issues or bug reporting: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.Windows/issues"
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Please refer to the AsBuiltReport.Microsoft.Windows github website for more detailed information about this project."
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Do not forget to update your report configuration file after each new release."
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Documentation: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.Windows"
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Issues or bug reporting: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.Windows/issues"
 
     Try {
         $InstalledVersion = Get-Module -ListAvailable -Name AsBuiltReport.Microsoft.Windows -ErrorAction SilentlyContinue | Sort-Object -Property Version -Descending | Select-Object -First 1 -ExpandProperty Version
 
         if ($InstalledVersion) {
-            Write-PScriboMessage -IsWarning "AsBuiltReport.Microsoft.Windows $($InstalledVersion.ToString()) is currently installed."
+            Write-PScriboMessage -Plugin "Module" -IsWarning "AsBuiltReport.Microsoft.Windows $($InstalledVersion.ToString()) is currently installed."
             $LatestVersion = Find-Module -Name AsBuiltReport.Microsoft.Windows -Repository PSGallery -ErrorAction SilentlyContinue | Select-Object -ExpandProperty Version
             if ($LatestVersion -gt $InstalledVersion) {
-                Write-PScriboMessage -IsWarning "AsBuiltReport.Microsoft.Windows $($LatestVersion.ToString()) is available."
-                Write-PScriboMessage -IsWarning "Run 'Update-Module -Name AsBuiltReport.Microsoft.Windows -Force' to install the latest version."
+                Write-PScriboMessage -Plugin "Module" -IsWarning "AsBuiltReport.Microsoft.Windows $($LatestVersion.ToString()) is available."
+                Write-PScriboMessage -Plugin "Module" -IsWarning "Run 'Update-Module -Name AsBuiltReport.Microsoft.Windows -Force' to install the latest version."
             }
         }
     } Catch {
@@ -235,7 +235,7 @@ function Invoke-AsBuiltReport.Microsoft.Windows {
                     if ($SMBShares) {
                         Section -Style Heading2 "File Server Configuration" {
                             Paragraph 'The following table details the File Server settings'
-                    BlankLine
+                            BlankLine
                             # SMB Server Configuration
                             Get-AbrWinSMBSummary
                             # SMB Server Network Interface
@@ -338,6 +338,58 @@ function Invoke-AsBuiltReport.Microsoft.Windows {
                     }
                 } else {
                     Write-PScriboMessage "No FailOver Cluster service detected. Disabling FailOver Cluster section"
+                }
+            }
+
+            if ($InfoLevel.SQLServer -ge 1 -and $OSType.Value -ne 'WorkStation') {
+                $Status = Invoke-Command -Session $TempPssSession -ScriptBlock { Get-Service 'MSSQL*' -ErrorAction SilentlyContinue }
+                if ($Status.Status -eq "Running") {
+                    try {
+                        if ($Options.SQLLogin) {
+                            $SQLCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $Options.SQLUserName, (ConvertTo-SecureString -Force $Options.SQLSecurePassword)
+                            $script:SQLServer = Connect-DbaInstance -SqlInstance $System -TrustServerCertificate -SqlCredential $SQLCredential
+                        } else {
+                            $script:SQLServer = Connect-DbaInstance -SqlInstance $System -TrustServerCertificate -SqlCredential $Credential
+                        }
+                        if ($SQLServer) {
+                            Section -Style Heading2 "SQL Server Configuration" {
+                                Paragraph "The following table details the SQL Server configuration from $($SQLServer.Name)."
+                                BlankLine
+                                # SQL Server Build Information
+                                Get-AbrWinSQLBuild
+                                # SQL Server Security Information
+                                Section -Style Heading3 "Security" {
+                                    Paragraph 'The following table details the SQL Server security settings'
+                                    BlankLine
+                                    # SQL Server Roles Information
+                                    Get-AbrWinSQLRole
+                                    # SQL Server Logins Information
+                                    Get-AbrWinSQLLogin
+                                }
+                                # SQL Server Database Information
+                                Get-AbrWinSQLDatabase
+                                # SQL Server Server Objects Information
+                                $BackupDevices = Get-AbrWinSQLBackupDevice
+                                if ($BackupDevices) {
+                                    Section -Style Heading3 "Server Objects" {
+                                        Paragraph 'The following table details the SQL Server server objects settings'
+                                        BlankLine
+                                        # SQL Server Backup Devices Information
+                                        $BackupDevices
+                                    }
+                                }
+                            }
+                            # Disconnect SQL Instance
+                            Write-PScriboMessage "Disconnecting SQL Instance"
+                            $SQLServer | Disconnect-DbaInstance | Out-Null
+                        } else {
+                            Write-PScriboMessage -IsWarning "Unable to connect to SQL Instance"
+                        }
+                    } catch {
+                        Write-PScriboMessage -IsWarning $_.Exception.Message
+                    }
+                } else {
+                    Write-PScriboMessage "No SQL Server service detected. Disabling SQL Server section"
                 }
             }