+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3] - 2024-03-06
+
+### Added
+
+- Add SQL Server support
+  - General Information
+  - Security
+    - Login
+    - Roles
+  - Databases
+    - User
+    - System
+  - Server Objects
+    - Backup Devices
+
 ## [0.5.2] - 2024-02-18
 
 ### Added